apps/api/pyproject.toml
[project]
name = "studybudd-api"
version = "0.1.0"
description = "StudyBudd API - FastAPI backend with PydanticAI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "pydantic-ai>=0.0.20",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.30.0",
    "greenlet>=3.0.0",
    "pgvector>=0.3.0",
    "alembic>=1.14.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "supabase>=2.10.0",
    "python-multipart>=0.0.9",
    "PyJWT>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["alembic*", "tests*"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.12"
strict = true
